July/Python/API_math_microservice/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
pika==1.3.2
prometheus-fastapi-instrumentator==7.1.0
cachetools==6.1.0
python-dotenv==1.1.1
kafka-python==2.2.15
